update and delete are finished

+++ blog/models/boards.py
@@ -312,6 +312,9 @@ def auto_delete_file_on_delete_for_blog(sender, instance=None, **kwargs):
             if field_type == 'FileField' or field_type == 'ImageField' or field_type == 'ImageSpecField':
                 origin_file = getattr(instance, field.name)
 
+                if origin_file == 'title_image':
+                    continue
+
                 if origin_file and os.path.isfile(origin_file.path):
                     os.remove(origin_file.path)
                     logger.debug('{} field file is deleted name of {} at model of {}'.format(field_type,origin_file,sender.__name__))
